Fixed query parameter storage.
Previously the http.Request object was augmented to store the query
parameters as a new field. Unfortunately this can (and will) break
compatibility with other middleware libraries.

Instead, based on feedback to the MicroCache middleware:
https://github.com/kevburnsjr/microcache/issues/10#issuecomment-534806688
the functionality has been updated to store the query parameters
directly in the HTTP context, and a helper function is provided to
extract them.